[CHANGE] [BRANCH/TAG] 0.1v Beta UnClean
Adding LicensePage and few theme changes

------------------------------

This commit is the last commit before making a branch or tag "0.1v Beta UnClean"

UnClean means the code includes a lot of comments or deprecated code,  and may include files that can/should be removed.

version 0.1v may not have a release, instead version 0.2v is the target after this commit, bugs that exist in this version, should be fixed in the 0.2v development.